+++ package.json
@@ -138,7 +138,7 @@
     "copy-webpack-plugin": "6.4.1",
     "coveralls": "3.1.0",
     "cross-env": "7.0.3",
-    "css-loader": "5.2.6",
+    "css-loader": "3.4.2",
     "cz-conventional-changelog": "3.3.0",
     "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "1.15.6",
