+++ packages/blockchain-wallet-v4-frontend/src/data/components/brokerage/actionTypes.ts
@@ -18,6 +18,8 @@ export const FETCH_BANK_TRANSFER_UPDATE = '@EVENT.FETCH_BANK_TRANSFER_UPDATE'
 export const FETCH_BANK_TRANSFER_UPDATE_LOADING = '@EVENT.FETCH_BANK_TRANSFER_UPDATE_LOADING'
 export const FETCH_BANK_TRANSFER_UPDATE_ERROR = '@EVENT.FETCH_BANK_TRANSFER_UPDATE_ERROR'
 
+export const FETCH_RECURRING_BUY_METHODS = '@EVENT.FETCH_RECURRING_BUY_METHODS'
+
 export const HANDLE_DEPOSIT_FIAT_CLICK = '@EVENT.HANDLE_DEPOSIT_FIAT_CLICK'
 export const HANDLE_WITHDRAW_CLICK = '@EVENT.HANDLE_WITHDRAW_CLICK'
 export const SET_ADD_BANK_STEP = '@EVENT.SET_BROKERAGE_STEP'
