+++ packages/blockchain-wallet-v4-frontend/src/data/components/brokerage/sagas.ts
@@ -412,12 +412,18 @@ export default ({ api, coreSagas, networks }: { api: APIType; coreSagas: any; ne
     }
   }
 
+  const fetchRBMethods = function* () {
+    const data: { eligibleMethods: SBPaymentTypes[] } = yield call(api.getRBPaymentMethods)
+    yield put(actions.form.change('recurringBuyScheduler', 'methods', data.eligibleMethods))
+  }
+
   return {
     createFiatDeposit,
     deleteSavedBank,
     fetchBankLinkCredentials,
     fetchBankTransferAccounts,
     fetchBankTransferUpdate,
+    fetchRBMethods,
     handleDepositFiatClick,
     handleMethodChange,
     handleWithdrawClick,
