+++ packages/blockchain-wallet-v4/package.json
@@ -95,6 +95,7 @@
     "scrypt-js": "3.0.1",
     "seedrandom": "2.4.3",
     "stellar-sdk": "6.0.0",
-    "unorm": "1.6.0"
+    "unorm": "1.6.0",
+    "uuid": "8.3.2"
   }
 }
